sentry - clean - move setupRaven to lib
