+++ app/scripts/lib/setupRaven.js
@@ -1,4 +1,4 @@
-const Raven = require('./vendor/raven.min.js')
+const Raven = require('../vendor/raven.min.js')
 const METAMASK_DEBUG = 'GULP_METAMASK_DEBUG'
 const PROD = 'https://3567c198f8a8412082d32655da2961d0@sentry.io/273505'
 const DEV = 'https://f59f3dd640d2429d9d0e2445a87ea8e1@sentry.io/273496'
