+++ app/scripts/popup.js
@@ -8,7 +8,7 @@ const extension = require('extensionizer')
 const ExtensionPlatform = require('./platforms/extension')
 const NotificationManager = require('./lib/notification-manager')
 const notificationManager = new NotificationManager()
-const setupRaven = require('./setupRaven')
+const setupRaven = require('./lib/setupRaven')
 
 // create platform global
 global.platform = new ExtensionPlatform()
